+++ container-engine-lib/lib/database_accessors/enclave_db/service_registration/repository_test.go
@@ -304,34 +304,12 @@ func getServiceRegistrationWithDataForTest(
 }
 
 func getServiceConfigForTest(t *testing.T, imageName string) *service.ServiceConfig {
-	serviceConfig, err := service.CreateServiceConfig(
-		imageName,
-		nil,
-		nil,
-		nil,
-		testPrivatePorts(t),
-		testPublicPorts(t),
-		[]string{"bin", "bash", "ls"},
-		[]string{"-l", "-a"},
-		testEnvVars(),
-		testFilesArtifactExpansion(),
-		testPersistentDirectory(),
-		500,
-		1024,
-		"IP-ADDRESS",
-		100,
-		512,
-		map[string]string{
-			"test-label-key":        "test-label-value",
-			"test-second-label-key": "test-second-label-value",
-		},
-		nil,
-		nil,
-		map[string]string{
-			"disktype": "ssd",
-		},
-		image_download_mode.ImageDownloadMode_Missing,
-	)
+	serviceConfig, err := service.CreateServiceConfig(imageName, nil, nil, nil, testPrivatePorts(t), testPublicPorts(t), []string{"bin", "bash", "ls"}, []string{"-l", "-a"}, testEnvVars(), testFilesArtifactExpansion(), testPersistentDirectory(), 500, 1024, "IP-ADDRESS", 100, 512, map[string]string{
+		"test-label-key":        "test-label-value",
+		"test-second-label-key": "test-second-label-value",
+	}, nil, nil, map[string]string{
+		"disktype": "ssd",
+	}, image_download_mode.ImageDownloadMode_Missing, true)
 	require.NoError(t, err)
 	return serviceConfig
 }