+++ core/server/api_container/server/startosis_engine/kurtosis_instruction/tasks/tasks_shared.go
@@ -50,6 +50,9 @@ const (
 	shellWrapperCommand = "/bin/sh"
 	noNameSet           = ""
 	uniqueNameGenErrStr = "error occurred while generating unique name for the file artifact"
+
+	//  enables init mode on containers; cleaning up any zombie processes
+	tiniEnabled = true
 )
 
 var runTailCommandToPreventContainerToStopOnCreating = []string{"tail", "-f", "/dev/null"}
@@ -271,34 +274,7 @@ func getServiceConfig(
 	filesArtifactExpansion *service_directory.FilesArtifactsExpansion,
 	envVars *map[string]string,
 ) (*service.ServiceConfig, error) {
-	serviceConfig, err := service.CreateServiceConfig(
-		maybeImageName,
-		maybeImageBuildSpec,
-		maybeImageRegistrySpec,
-		maybeNixBuildSpec,
-		nil,
-		nil,
-		// This make sure that the container does not stop as soon as it starts
-		// This only is needed for kubernetes at the moment
-		// TODO: Instead of creating a service and running exec commands
-		//  we could probably run the command as an entrypoint and retrieve the results as soon as the
-		//  command is completed
-		runTailCommandToPreventContainerToStopOnCreating,
-		nil,
-		*envVars,
-		filesArtifactExpansion,
-		nil,
-		0,
-		0,
-		service_config.DefaultPrivateIPAddrPlaceholder,
-		0,
-		0,
-		map[string]string{},
-		nil,
-		nil,
-		map[string]string{},
-		image_download_mode.ImageDownloadMode_Missing,
-	)
+	serviceConfig, err := service.CreateServiceConfig(maybeImageName, maybeImageBuildSpec, maybeImageRegistrySpec, maybeNixBuildSpec, nil, nil, runTailCommandToPreventContainerToStopOnCreating, nil, *envVars, filesArtifactExpansion, nil, 0, 0, service_config.DefaultPrivateIPAddrPlaceholder, 0, 0, map[string]string{}, nil, nil, map[string]string{}, image_download_mode.ImageDownloadMode_Missing, tiniEnabled)
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred creating service config")
 	}