+++ engine/server/engine/centralized_logs/client_implementations/persistent_volume/log_file_manager/log_file_manager_test.go
@@ -44,7 +44,7 @@ func TestRemoveLogsBeyondRetentionPeriod(t *testing.T) {
 	_, _ = mockFs.Create(week1filepath)
 	_, _ = mockFs.Create(week2filepath)
 
-	logFileManager := NewLogFileManager(mockKurtosisBackend, mockFs, mockTime)
+	logFileManager := NewLogFileManager(mockKurtosisBackend, mockFs, mockTime, 5)
 	logFileManager.RemoveLogsBeyondRetentionPeriod() // should remove week 49 logs
 
 	_, err := mockFs.Stat(week49filepath)
@@ -69,7 +69,7 @@ func TestRemoveEnclaveLogs(t *testing.T) {
 	_, _ = mockFs.Create(week52filepath)
 	_, _ = mockFs.Create(week52filepathDiffService)
 
-	logFileManager := NewLogFileManager(mockKurtosisBackend, mockFs, mockTime)
+	logFileManager := NewLogFileManager(mockKurtosisBackend, mockFs, mockTime, 5)
 	err := logFileManager.RemoveEnclaveLogs(testEnclaveUuid) // should remove only all log files for enclave one
 
 	require.NoError(t, err)
@@ -107,7 +107,7 @@ func TestRemoveAllLogs(t *testing.T) {
 	_, _ = mockFs.Create(week52filepath)
 	_, _ = mockFs.Create(week52filepathDiffService)
 
-	logFileManager := NewLogFileManager(mockKurtosisBackend, mockFs, mockTime)
+	logFileManager := NewLogFileManager(mockKurtosisBackend, mockFs, mockTime, 5)
 	err := logFileManager.RemoveAllLogs()
 
 	require.NoError(t, err)
@@ -165,7 +165,7 @@ func TestCreateLogFiles(t *testing.T) {
 	expectedServiceNameFilePath := getFilepathStr(2022, 52, testEnclaveUuid, testUserService1Name)
 	expectedServiceShortUuidFilePath := getFilepathStr(2022, 52, testEnclaveUuid, uuid_generator.ShortenedUUIDString(testUserService1Uuid))
 
-	logFileManager := NewLogFileManager(mockKurtosisBackend, mockFs, mockTime)
+	logFileManager := NewLogFileManager(mockKurtosisBackend, mockFs, mockTime, 5)
 	err := logFileManager.CreateLogFiles(ctx)
 	require.NoError(t, err)