+++ engine/server/engine/centralized_logs/client_implementations/persistent_volume/volume_consts/consts.go
@@ -17,8 +17,8 @@ const (
 
 	EndOfJsonLine = "}"
 
-	// promise to keep 4 weeks of logs for users, but store an additional week for safety
-	LogRetentionPeriodInWeeks = 5
+	// promise to keep 1 weeks of logs for users
+	LogRetentionPeriodInWeeks = 1
 
 	RemoveLogsWaitHours = 6 * time.Hour